based on: http://www.freeallegiance.org/forums/index.php?showtopic=66604&view=findpost&p=640554
Looking at Juckto's explaination it struck me that it's not 'sunday' that's important, but 'squad game peak' that we want to call out. In that light, I am treating 19:00 UTC on Sunday as midnight, and adjusting the graph labels accordingly so that if Juckto looks at the graph at UTC+13:00, he's going to see that the best time to login is 9:00am on monday local time, which I think is optimal. I also added a timezone dropdown that auto sets to your local time, and lets you client side shift the graph and the squad game time on change. (BackTrak)